+++ src/js/components/MultiSelect/MultiSelect.js
@@ -10,12 +10,7 @@ import React, {
 } from 'react';
 import styled, { ThemeContext } from 'styled-components';
 
-import {
-  controlBorderStyle,
-  normalizeColor,
-  useKeyboard,
-  useForwardedRef,
-} from '../../utils';
+import { controlBorderStyle, useKeyboard, useForwardedRef } from '../../utils';
 import { defaultProps } from '../../default-props';
 
 import { Box } from '../Box';
@@ -24,26 +19,25 @@ import { CheckBox } from '../CheckBox';
 import { DropButton } from '../DropButton';
 import { Keyboard } from '../Keyboard';
 import { FormContext } from '../Form/FormContext';
-import { TextInput } from '../TextInput';
 import { Text } from '../Text';
 
 import { MultiSelectContainer } from './MultiSelectContainer';
-import { applyKey, HiddenInput } from '../Select/utils';
+import {
+  applyKey,
+  HiddenInput,
+  SelectTextInput,
+  StyledSelectDropButton,
+  calcValuedValue,
+  changeEvent,
+  getSelectIcon,
+  checkDisabled,
+  getIconColor,
+  getOptionValue,
+  getOptionLabel,
+  getDisplayLabelKey,
+} from '../Select/utils';
 import { MessageContext } from '../../contexts/MessageContext';
 import { MultiSelectPropTypes } from './propTypes';
-
-const SelectTextInput = styled(TextInput)`
-  cursor: ${(props) => (props.defaultCursor ? 'default' : 'pointer')};
-`;
-
-const StyledSelectDropButton = styled(DropButton)`
-  ${(props) => !props.callerPlain && controlBorderStyle};
-  ${(props) =>
-    props.theme.select &&
-    props.theme.select.control &&
-    props.theme.select.control.extend};
-  ${(props) => props.open && props.theme.select.control.open};
-`;
 
 const StyledSelectBox = styled(Box)`
   ${(props) => !props.callerPlain && controlBorderStyle};
@@ -114,10 +108,6 @@ const MultiSelect = forwardRef(
     const selectBoxRef = useRef();
     const [showA11yDiv, setShowA11yDiv] = useState(false);
     const dropButtonRef = useForwardedRef(ref);
-
-    // Determine if the Select is opened with the keyboard. If so,
-    // focus should be set on the first option when the drop opens
-    // see set initial focus code in SelectContainer.js
     const usingKeyboard = useKeyboard();
 
     // value is used for what we receive in valueProp and the basis for
@@ -132,22 +122,11 @@ const MultiSelect = forwardRef(
       value: valueProp,
       initialValue: defaultValue || '',
     });
-    // valuedValue is the value mapped with any valueKey applied
-    // When the options array contains objects, this property indicates how
-    // to retrieve the value of each option.
-    // If a string is provided, it is used as the key to retrieve a
-    // property of an option object.
-    // If a function is provided, it is called with the option and should
-    // return the value.
-    // If reduce is true, this value will be used for the 'value'
-    // delivered via 'onChange'.
-    const valuedValue = useMemo(() => {
-      if (Array.isArray(value))
-        return value.map((v) =>
-          valueKey && valueKey.reduce ? v : applyKey(v, valueKey),
-        );
-      return valueKey && valueKey.reduce ? value : applyKey(value, valueKey);
-    }, [value, valueKey]);
+
+    const valuedValue = useMemo(
+      () => calcValuedValue(value, valueKey),
+      [value, valueKey],
+    );
     // search input value
     const [search, setSearch] = useState();
     // All select option indices and values
@@ -189,19 +168,10 @@ const MultiSelect = forwardRef(
       setSearch();
     }, [onClose]);
 
-    const triggerChangeEvent = useCallback((nextValue) => {
-      // Calling set value function directly on input because React library
-      // overrides setter `event.target.value =` and loses original event
-      // target fidelity.
-      // https://stackoverflow.com/a/46012210
-      const nativeInputValueSetter = Object.getOwnPropertyDescriptor(
-        window.HTMLInputElement.prototype,
-        'value',
-      ).set;
-      nativeInputValueSetter.call(inputRef.current, nextValue);
-      const event = new Event('input', { bubbles: true });
-      inputRef.current.dispatchEvent(event);
-    }, []);
+    const triggerChangeEvent = useCallback(
+      (nextValue) => changeEvent(inputRef, nextValue),
+      [],
+    );
 
     const onSelectChange = useCallback(
       (event, { option, value: nextValue }) => {
@@ -243,48 +213,22 @@ const MultiSelect = forwardRef(
       [onChange, setValue, triggerChangeEvent],
     );
 
-    let SelectIcon;
-    switch (icon) {
-      case false:
-        break;
-      case true:
-      case undefined:
-        SelectIcon =
-          open && theme.select.icons.up
-            ? theme.select.icons.up
-            : theme.select.icons.down;
-        break;
-      default:
-        SelectIcon = icon;
-    }
+    const SelectIcon = getSelectIcon(icon, theme, open);
+
+    const optionLabel = useCallback(
+      (index) => getOptionLabel(index, allOptions, labelKey),
+      [labelKey, allOptions],
+    );
 
     const optionValue = useCallback(
-      (index) => applyKey(allOptions[index], valueKey),
+      (index) => getOptionValue(index, allOptions, valueKey),
       [allOptions, valueKey],
     );
 
     const isDisabled = useCallback(
-      (index) => {
-        const option = allOptions[index];
-        let result;
-        if (disabledKey) {
-          result = applyKey(option, disabledKey);
-        } else if (Array.isArray(disabled)) {
-          if (typeof disabled[0] === 'number') {
-            result = disabled.indexOf(index) !== -1;
-          } else {
-            const optionVal = optionValue(index);
-            result = disabled.indexOf(optionVal) !== -1;
-          }
-        }
-        return result;
-      },
-      [disabled, disabledKey, allOptions, optionValue],
-    );
-
-    const optionLabel = useCallback(
-      (index) => applyKey(allOptions[index], labelKey),
-      [labelKey, allOptions],
+      (index) =>
+        checkDisabled(index, disabled, disabledKey, allOptions, valueKey),
+      [disabled, disabledKey, allOptions, valueKey],
     );
 
     // element to show, trumps inputValue
@@ -452,7 +396,6 @@ const MultiSelect = forwardRef(
             )}
           </>
         );
-        // }
       }
       return undefined;
     }, [
@@ -481,22 +424,16 @@ const MultiSelect = forwardRef(
       }
     }, [showA11yDiv]);
 
-    // if labelKey is a function and valueLabel is not defined
-    // we should use the labelKey function to display the
-    // selected value
-    const displayLabelKey = useMemo(() => {
-      const optionLabelKey = applyKey(
-        allOptions[optionIndexesInValue[0]],
-        labelKey,
-      );
-      if (
-        !selectValue &&
-        optionIndexesInValue.length === 1 &&
-        typeof optionLabelKey === 'object'
-      )
-        return optionLabelKey;
-      return undefined;
-    }, [labelKey, allOptions, optionIndexesInValue, selectValue]);
+    const displayLabelKey = useMemo(
+      () =>
+        getDisplayLabelKey(
+          labelKey,
+          allOptions,
+          optionIndexesInValue,
+          selectValue,
+        ),
+      [labelKey, allOptions, optionIndexesInValue, selectValue],
+    );
 
     // text to show
     // When the options array contains objects, this property indicates how
@@ -525,10 +462,7 @@ const MultiSelect = forwardRef(
       selectValue,
     ]);
 
-    const iconColor = normalizeColor(
-      theme.select.icons.color || 'control',
-      theme,
-    );
+    const iconColor = getIconColor(theme);
 
     const dropContent = (
       <MultiSelectContainer
